Remove extra layers
According to the given structure in the paper, there is an extra convolutional layer in green block (inp_res). Also, before the final layer there is another extra layer called Input_Dec_Output. It is probably because the article is a little vague, but I checked this architecture with the paper of the winner of BraTS 2019. The winner of BraTS 2019 has used the same architecture, but they have explained it much clear. This is their paper : https://link.springer.com/chapter/10.1007/978-3-030-46640-4_22